# app/docs/swagger.yaml
definitions:
...
  models.CategoryCreateBody:
    properties:
      description:
        type: string
      image:
        type: string
      name:
        type: string
    type: object
  models.CategoryDeleteBody:
    properties:
      category_id:
        items:
          type: string
        type: array
    type: object
  models.CategoryInfo:
    properties:
      category_id:
        type: string
      created:
        type: string
      creator_id:
        type: integer
      description:
        type: string
      image:
        type: string
      name:
        type: string
      updated:
        type: string
    type: object
  models.CategoryUpdateBody:
    properties:
      category_id:
        type: string
      description:
        type: string
      image:
        type: string
      name:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      success:
        type: boolean
    type: object
...
  models.SuccessResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact: {}
paths:
  /admin/categories/create:
    post:
      consumes:
      - application/json
      description: Endpoint to create product category
      parameters:
      - description: request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CategoryCreateBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Created product category
      tags:
      - Categories
  /admin/categories/delete:
    delete:
      consumes:
      - application/json
      description: Endpoint to delete category
      parameters:
      - description: request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CategoryDeleteBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Delete category
      tags:
      - Categories
  /admin/categories/info:
    get:
      consumes:
      - application/json
      description: Endpoint to get information about category by id
      parameters:
      - description: category id
        in: query
        name: category_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CategoryInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Get category info by id
      tags:
      - Categories
  /admin/categories/list:
    get:
      consumes:
      - application/json
      description: Endpoint to get list of categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CategoryInfo'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Get categories list
      tags:
      - Categories
  /admin/categories/update:
    patch:
      consumes:
      - application/json
      description: Endpoint to update category
      parameters:
      - description: request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CategoryUpdateBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Update category
      tags:
      - Categories
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"